refactor(core): Move methods from WorkflowHelpers to WorkflowService (no-changelog)
