+++ packages/cli/test/integration/executions.controller.test.ts
@@ -1,10 +1,13 @@
 import type { User } from '@db/entities/User';
+import { Push } from '@/push';
 import { createSuccessfulExecution, getAllExecutions } from './shared/db/executions';
 import { createOwner } from './shared/db/users';
 import { createWorkflow } from './shared/db/workflows';
 import * as testDb from './shared/testDb';
 import { setupTestServer } from './shared/utils';
+import { mockInstance } from '../shared/mocking';
 
+mockInstance(Push);
 let testServer = setupTestServer({ endpointGroups: ['executions'] });
 
 let owner: User;
