+++ packages/cli/test/integration/workflows/workflows.controller.test.ts
@@ -1,27 +1,28 @@
+import Container from 'typedi';
 import type { SuperAgentTest } from 'supertest';
+import { v4 as uuid } from 'uuid';
 import type { INode, IPinData } from 'n8n-workflow';
+
 import * as UserManagementHelpers from '@/UserManagement/UserManagementHelper';
 import type { User } from '@db/entities/User';
-import { v4 as uuid } from 'uuid';
+import { WorkflowRepository } from '@db/repositories/workflow.repository';
+import type { WorkflowEntity } from '@db/entities/WorkflowEntity';
 import { RoleService } from '@/services/role.service';
-import Container from 'typedi';
 import type { ListQuery } from '@/requests';
 import { WorkflowHistoryRepository } from '@db/repositories/workflowHistory.repository';
 import { ActiveWorkflowRunner } from '@/ActiveWorkflowRunner';
-
-import { mockInstance } from '../shared/mocking';
-import * as utils from './shared/utils/';
-import * as testDb from './shared/testDb';
-import { makeWorkflow, MOCK_PINDATA } from './shared/utils/';
-import { randomCredentialPayload } from './shared/random';
-import { saveCredential } from './shared/db/credentials';
-import { createOwner } from './shared/db/users';
-import { createWorkflow } from './shared/db/workflows';
-import { createTag } from './shared/db/tags';
 import { Push } from '@/push';
 import { EnterpriseWorkflowService } from '@/workflows/workflow.service.ee';
-import { WorkflowRepository } from '@/databases/repositories/workflow.repository';
-import type { WorkflowEntity } from '@/databases/entities/WorkflowEntity';
+
+import { mockInstance } from '../../shared/mocking';
+import * as utils from '../shared/utils/';
+import * as testDb from '../shared/testDb';
+import { makeWorkflow, MOCK_PINDATA } from '../shared/utils/';
+import { randomCredentialPayload } from '../shared/random';
+import { saveCredential } from '../shared/db/credentials';
+import { createOwner } from '../shared/db/users';
+import { createWorkflow } from '../shared/db/workflows';
+import { createTag } from '../shared/db/tags';
 
 let owner: User;
 let authOwnerAgent: SuperAgentTest;
